+++ shared/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.0.60] - 2023-06-06
 
 - Add place-holders for servicer token usage support
+- Add a TODO comment for supporting different relay payload types in core types
 
 ## [0.0.0.59] - 2023-06-01